+++ src/handlers/filesHandlers.ts
@@ -1,13 +1,17 @@
 import { requireEnv } from 'require-env-variable';
 import { HTTP_STATUS } from '../constants';
 import { isAuthContext, isWritableFile } from '../types';
-import { createFile } from '../database/operations';
+import { createFile, loadOrCreateS3Bucket } from '../database/operations';
 import { db } from '../database';
 import { InputValidationError, FailedMiddlewareError } from '../errors';
 import { generatePresignedPost } from '../s3';
 import type { Request, Response } from 'express';
 
-const { S3_BUCKET, S3_REGION } = requireEnv('S3_BUCKET', 'S3_REGION');
+const { S3_BUCKET, S3_REGION, S3_ENDPOINT } = requireEnv(
+	'S3_BUCKET',
+	'S3_REGION',
+	'S3_ENDPOINT',
+);
 
 const postFile = async (req: Request, res: Response): Promise<void> => {
 	if (!isAuthContext(req)) {
@@ -23,12 +27,17 @@ const postFile = async (req: Request, res: Response): Promise<void> => {
 	}
 
 	const { name, mimeType, size } = body;
+	const s3Bucket = await loadOrCreateS3Bucket(db, req, {
+		name: S3_BUCKET,
+		region: S3_REGION,
+		endpoint: S3_ENDPOINT,
+	});
+
 	const file = await createFile(db, req, {
 		name,
 		mimeType,
 		size,
-		bucketName: S3_BUCKET,
-		bucketRegion: S3_REGION,
+		s3BucketName: s3Bucket.name,
 	});
 	const presignedPost = await generatePresignedPost(
 		file.storageKey,